🚀 [#1] Feature: Initiate e2eGenericCrudTest tool, add image field to migrations and product model, and create seeders for products 🧪

+++ code/db/migrations/20240623214858-create-default-admin.js
@@ -14,6 +14,6 @@ module.exports = {
 
   async down () {
     const user =await userService.findByEmail('admin@local.com');
-    await userService.delete(user.id);
+    if(user) await userService.delete(user.id);
   }
 };
